requirements-dev.txt
pyyaml==6.0.3
pynacl>=1.5.0
fastavro>=1.9.0
rdflib>=7,<8
pyshacl>=0.28,<1
